requirements.txt
fastapi
uvicorn
python-dotenv
openai
langchain
langchain-community
langchain-openai
serpapi
google-search-results    # ← required by SerpAPIWrapper
python-docx
requests
sendgrid